+++ storybook-static/assets/preview-21c47d67.js
@@ -1,0 +1,2 @@
+import{_ as e}from"./iframe-934475b8.js";import"../sb-preview/runtime.js";var a={docs:{renderer:async()=>{let{DocsRenderer:r}=await e(()=>import("./DocsRenderer-3PUGWF3O-ee789181.js"),["./DocsRenderer-3PUGWF3O-ee789181.js","./chunk-S4VUQJ4A-01a18434.js","./iframe-934475b8.js","./index-d8ade8ea.js","./_commonjsHelpers-042e6b4d.js","./index-d475d2ea.js","./index-d37d4223.js","./index-6ba087f2.js","./floating-ui.dom-c3a31ea6.js","./index-356e4a49.js"],import.meta.url);return new r}}};export{a as parameters};
+//# sourceMappingURL=preview-21c47d67.js.map